scripts/requirements-gmail.txt
# Medallion: Bronze | Mutation: 0% | Component: Gmail Agent Dependencies
google-api-python-client==2.155.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
